# liftflexSwagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "v2",
        "description": "",
        "version": "1"
    },
    "basePath": "/Liftflex_API/rest/v2",
    "schemes": [
        "https"
    ],
    "paths": {
        "/GetBoletoByCprf": {
            "get": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "GetBoletoByCprf",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Cprf",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TenantId",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "x-example": 0,
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Retorno"
                        }
                    }
                }
            }
        },
        "/GetBoletoById": {
            "get": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "GetBoletoById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Id",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TenandId",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "x-example": 0,
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/RetornoBoleto"
                        }
                    }
                }
            }
        },
...
        "/GetContasReceberById": {
            "get": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "GetContasReceberById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Key",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "ContasReceberId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/GetFinanceiroCRStruct"
                        }
                    }
                }
            }
        },
        "/GetContasReceberByVencimento": {
            "get": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "GetContasReceberByVencimento",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Key",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "DataInicial",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "DataFinal",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/GetFinanceiroCRStruct"
                        }
                    }
                }
            }
        },
        "/GetContatoCliente": {
            "get": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "GetContatoCliente",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Key",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "ClienteId",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "x-example": 1234567891234567,
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/ContatosListBooleanTextLongIntegerRecord"
                        }
                    }
                }
            }
        },
...
        "/UpdateWhatsAppBoletoMessage": {
            "post": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "UpdateWhatsAppBoletoMessage",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "BoletoId",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "x-example": 0,
                        "format": "int32"
                    },
                    {
                        "name": "TenantId",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "x-example": 0,
                        "format": "int32"
                    },
                    {
                        "name": "WhatsAppBoletoResponse",
                        "in": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/WhatsAppBoletoResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/ValidarCprfTelefone": {
            "get": {
                "tags": [
                    "v2"
                ],
                "description": "",
                "operationId": "ValidarCprfTelefone",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "TenantId",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "x-example": 0,
                        "format": "int32"
                    },
                    {
                        "name": "Cprf",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "Telefone",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/ValidarCprfTelefoneResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "BoletoResponse": {
            "description": "",
            "type": "object",
            "properties": {
                "Identificador": {
                    "type": "integer",
                    "example": 1234567891234567,
                    "format": "int64"
                },
                "Vencimento": {
                    "type": "string",
                    "default": ""
                },
                "Valor": {
                    "type": "number",
                    "example": 0.1
                },
                "NossoNumero": {
                    "type": "string",
                    "default": ""
                },
                "LinhaDigitavel": {
                    "type": "string",
                    "default": ""
                },
                "Binario": {
                    "type": "string"
                }
            }
        },
...
        "Contatos": {
            "description": "",
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Nome": {
                    "type": "string",
                    "default": ""
                },
                "DataNascimento": {
                    "type": "string",
                    "default": ""
                },
                "Rascunho": {
                    "type": "string",
                    "default": ""
                },
                "ContatosTelefone": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ContatosTelefone"
                    }
                },
                "ContatosEmails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ContatosEmail"
                    }
                }
            }
        },
        "ContatosEmail": {
            "description": "",
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Email": {
                    "type": "string",
                    "default": ""
                },
                "Rascunho": {
                    "type": "string",
                    "default": ""
                },
                "EventoEmailTable": {
                    "$ref": "#/definitions/EventoEmail"
                }
            }
        },
        "ContatosListBooleanTextLongIntegerRecord": {
            "description": "",
            "type": "object",
            "properties": {
                "Mensagem": {
                    "type": "string",
                    "default": ""
                },
                "Total": {
                    "type": "integer",
                    "example": 1234567891234567,
                    "format": "int64"
                },
                "IsSucesso": {
                    "type": "boolean",
                    "example": false
                },
                "Contatos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Contatos"
                    }
                }
            }
        },
        "ContatosTelefone": {
            "description": "",
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Telefone": {
                    "type": "string",
                    "default": ""
                },
                "Padrao": {
                    "type": "string",
                    "default": ""
                },
                "EventoTelefone": {
                    "$ref": "#/definitions/EventoTelefone"
                }
            }
        },
...
        "EventoEmail": {
            "description": "",
            "type": "object",
            "properties": {
                "EventoEmailId": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Label": {
                    "type": "string",
                    "default": ""
                }
            }
        },
        "EventoTelefone": {
            "description": "",
            "type": "object",
            "properties": {
                "EvendoTelefoneId": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Label": {
                    "type": "string",
                    "default": ""
                }
            }
        },
...
        "FinanceiroCRConta": {
            "description": "",
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Situacao": {
                    "type": "string",
                    "default": ""
                },
                "TipoOperacao": {
                    "type": "string",
                    "default": ""
                },
                "IdOperacao": {
                    "type": "string",
                    "default": ""
                },
                "Documento": {
                    "type": "string",
                    "default": ""
                },
                "DataHoraCadastro": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2014-12-31T23:59:59.938Z"
                },
                "DataHoraManutencao": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2014-12-31T23:59:59.938Z"
                },
                "Historico": {
                    "type": "string",
                    "default": ""
                },
                "FormaPagamento": {
                    "type": "string",
                    "default": ""
                },
                "TotalParcelas": {
                    "type": "string",
                    "default": ""
                },
                "NumeroParcela": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "SubTotal": {
                    "type": "number",
                    "example": 0.1
                },
                "DescontoTotal": {
                    "type": "number",
                    "example": 0.1
                },
                "JurosTotal": {
                    "type": "number",
                    "example": 0.1
                },
                "MultaTotal": {
                    "type": "number",
                    "example": 0.1
                },
                "Total": {
                    "type": "number",
                    "example": 0.1
                },
                "DataVencimento": {
                    "type": "string",
                    "format": "date",
                    "example": "2014-12-31"
                },
                "ValorPago": {
                    "type": "number",
                    "example": 0.1
                },
                "DataPagamento": {
                    "type": "string",
                    "format": "date",
                    "example": "2014-12-31"
                },
                "ImpostoRetido": {
                    "type": "number",
                    "example": 0.1
                }
            },
            "required": [
                "TipoOperacao",
                "DataHoraCadastro",
                "DataHoraManutencao",
                "Historico",
                "Parcelas",
                "SubTotal",
                "DescontoTotal",
                "JurosTotal",
                "MultaTotal",
                "Total",
                "DataVencimento",
                "ValorPago",
                "DataPagamento",
                "RetencaoValorTotal"
            ]
        },
        "FinanceiroCRPagador": {
            "description": "",
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32"
                },
                "Ativo": {
                    "type": "boolean",
                    "example": false
                },
                "Nome": {
                    "type": "string",
                    "default": ""
                },
                "NomeFantasia": {
                    "type": "string",
                    "default": ""
                },
                "Tipo": {
                    "type": "string",
                    "default": ""
                },
                "Cnpj": {
                    "type": "string",
                    "default": ""
                },
                "Cpf": {
                    "type": "string",
                    "default": ""
                }
            }
        },
        "FinanceiroCRRateioStruct": {
            "description": "",
            "type": "object",
            "properties": {
                "Id": {
                    "type": "integer",
                    "example": 0,
                    "format": "int32",
                    "description": "Id que identifica o rateio"
                },
                "PlanoContaOrdenador": {
                    "type": "string",
                    "default": ""
                },
                "PlanoContaNome": {
                    "type": "string",
                    "default": ""
                },
                "CentroCusto": {
                    "type": "string",
                    "default": ""
                },
                "Valor": {
                    "type": "number",
                    "example": 0.1
                },
                "ValorPago": {
                    "type": "number",
                    "example": 0.1
                },
                "Desconto": {
                    "type": "number",
                    "example": 0.1
                },
                "Multa": {
                    "type": "number",
                    "example": 0.1
                },
                "Juros": {
                    "type": "number",
                    "example": 0.1
                }
            },
            "required": [
                "ValorPago",
                "Desconto",
                "Multa",
                "Juros"
            ]
        },
        "FinanceiroCRStruct": {
            "description": "",
            "type": "object",
            "properties": {
                "Pagador": {
                    "$ref": "#/definitions/FinanceiroCRPagador"
                },
                "Conta": {
                    "$ref": "#/definitions/FinanceiroCRConta"
                },
                "Rateio": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FinanceiroCRRateioStruct"
                    }
                }
            }
        },
...
        "GetFinanceiroCRStruct": {
            "description": "",
            "type": "object",
            "properties": {
                "Resultado": {
                    "type": "boolean",
                    "example": false
                },
                "ResultadoMsg": {
                    "type": "string",
                    "default": ""
                },
                "Dados": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FinanceiroCRStruct"
                    }
                }
            },
            "required": [
                "Resultado",
                "ResultadoMsg",
                "Dados"
            ]
        },
...
        "Retorno": {
            "description": "",
            "type": "object",
            "properties": {
                "Resultado": {
                    "type": "boolean",
                    "example": false
                },
                "Mensagem": {
                    "type": "string",
                    "default": ""
                },
                "Boletos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/BoletoResponse"
                    }
                }
            }
        },
        "RetornoBoleto": {
            "description": "",
            "type": "object",
            "properties": {
                "Resultado": {
                    "type": "boolean",
                    "example": false
                },
                "Mensagem": {
                    "type": "string",
                    "default": ""
                },
                "Boleto": {
                    "$ref": "#/definitions/BoletoResponse"
                }
            }
        },
...
        "ValidarCprfTelefoneResponse": {
            "description": "",
            "type": "object",
            "properties": {
                "CprfValido": {
                    "type": "boolean",
                    "example": false
                },
                "TelefoneValido": {
                    "type": "boolean",
                    "example": false
                },
                "Mensagem": {
                    "type": "string",
                    "default": ""
                },
                "CodigoErro": {
                    "type": "string",
                    "default": ""
                }
            }
        },
        "WhatsAppBoletoResponse": {
            "description": "",
            "type": "object",
            "properties": {
                "Status": {
                    "type": "boolean",
                    "example": false
                },
                "Message": {
                    "type": "string",
                    "default": ""
                }
            },
            "required": [
                "Status"
            ]
        }
    },
...
}